Fix gallery app crash when printing images

+++ legacy/core/src/main/java/com/fsck/k9/provider/AttachmentTempFileProvider.java
@@ -5,22 +5,32 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.List;
 import java.util.Locale;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
+import android.content.pm.ProviderInfo;
+import android.database.Cursor;
+import android.database.MatrixCursor;
 import android.net.Uri;
 import android.os.AsyncTask;
+import android.os.ParcelFileDescriptor;
+import android.provider.OpenableColumns;
+
 import androidx.annotation.MainThread;
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.annotation.WorkerThread;
 import androidx.core.content.ContextCompat;
 import androidx.core.content.FileProvider;
-import timber.log.Timber;
+import java.io.FileNotFoundException;
 
 import com.fsck.k9.K9;
+import timber.log.Timber;
+
 import okio.ByteString;
 import org.apache.commons.io.IOUtils;
 
@@ -37,31 +47,48 @@ public class AttachmentTempFileProvider extends FileProvider {
 
     @Override
     public boolean onCreate() {
-        String packageName = getContext().getPackageName();
-        AUTHORITY = packageName + ".tempfileprovider";
-        return true;
+        return super.onCreate();
+    }
+
+    @Override
+    public void attachInfo(@NonNull Context context, @NonNull ProviderInfo info) {
+        AUTHORITY = info.authority;
+        super.attachInfo(context, info);
+    }
+
+    private static synchronized String getAuthority(Context context) {
+        if (AUTHORITY == null) {
+            AUTHORITY = context.getPackageName() + ".tempfileprovider";
+        }
+        return AUTHORITY;
     }
 
     @WorkerThread
     public static Uri createTempUriForContentUri(Context context, Uri uri, String displayName) throws IOException {
         Context applicationContext = context.getApplicationContext();
 
-        File tempFile = getTempFileForUri(uri, applicationContext);
+        String tempFilename = getTempFilenameForUri(uri);
+        if (displayName != null) {
+            int lastDot = displayName.lastIndexOf('.');
+            if (lastDot >= 0) {
+                String extension = displayName.substring(lastDot).toLowerCase(Locale.ENGLISH);
+                // Basic validation of extension
+                if (extension.length() > 1 && extension.length() < 10 && extension.indexOf('/') == -1) {
+                    tempFilename += extension;
+                }
+            }
+        }
+
+        File tempDirectory = getTempFileDirectory(applicationContext);
+        File tempFile = new File(tempDirectory, tempFilename);
+
         writeUriContentToTempFileIfNotExists(context, uri, tempFile);
-        Uri tempFileUri = FileProvider.getUriForFile(context, AUTHORITY, tempFile, displayName);
+
+        Uri tempFileUri = FileProvider.getUriForFile(context, getAuthority(applicationContext), tempFile);
 
         registerFileCleanupReceiver(applicationContext);
 
         return tempFileUri;
-    }
-
-    @NonNull
-    private static File getTempFileForUri(Uri uri, Context context) {
-        Context applicationContext = context.getApplicationContext();
-
-        String tempFilename = getTempFilenameForUri(uri);
-        File tempDirectory = getTempFileDirectory(applicationContext);
-        return new File(tempDirectory, tempFilename);
     }
 
     private static String getTempFilenameForUri(Uri uri) {
@@ -75,24 +102,19 @@ public class AttachmentTempFileProvider extends FileProvider {
                 return;
             }
 
-            FileOutputStream outputStream = new FileOutputStream(tempFile);
-            InputStream inputStream = context.getContentResolver().openInputStream(uri);
-            if (inputStream == null) {
-                throw new IOException("Failed to resolve content at uri: " + uri);
+            try (FileOutputStream outputStream = new FileOutputStream(tempFile);
+                 InputStream inputStream = context.getContentResolver().openInputStream(uri)) {
+                if (inputStream == null) {
+                    throw new IOException("Failed to resolve content at uri: " + uri);
+                }
+                IOUtils.copy(inputStream, outputStream);
             }
-            IOUtils.copy(inputStream, outputStream);
-
-            outputStream.close();
-            IOUtils.closeQuietly(inputStream);
         }
     }
 
     public static Uri getMimeTypeUri(Uri contentUri, String mimeType) {
-        if (!AUTHORITY.equals(contentUri.getAuthority())) {
-            throw new IllegalArgumentException("Can only call this method for URIs within this authority!");
-        }
         if (contentUri.getQueryParameter("mime_type") != null) {
-            throw new IllegalArgumentException("Can only call this method for not yet typed URIs!");
+            return contentUri;
         }
         return contentUri.buildUpon().appendQueryParameter("mime_type", mimeType).build();
     }
@@ -101,13 +123,16 @@ public class AttachmentTempFileProvider extends FileProvider {
         File tempDirectory = getTempFileDirectory(context);
         boolean allFilesDeleted = true;
         long deletionThreshold = System.currentTimeMillis() - FILE_DELETE_THRESHOLD_MILLISECONDS;
-        for (File tempFile : tempDirectory.listFiles()) {
+        File[] files = tempDirectory.listFiles();
+        if (files == null) {
+            return true;
+        }
+        for (File tempFile : files) {
             long lastModified = tempFile.lastModified();
             if (lastModified < deletionThreshold) {
                 boolean fileDeleted = tempFile.delete();
                 if (!fileDeleted) {
                     Timber.e("Failed to delete temporary file");
-                    // TODO really do this? might cause our service to stay up indefinitely if a file can't be deleted
                     allFilesDeleted = false;
                 }
             } else {
@@ -136,13 +161,85 @@ public class AttachmentTempFileProvider extends FileProvider {
 
 
     @Override
-    public String getType(Uri uri) {
-        return uri.getQueryParameter("mime_type");
+    public String getType(@NonNull Uri uri) {
+        String mimeType = uri.getQueryParameter("mime_type");
+        if (mimeType != null) {
+            return mimeType;
+        }
+
+        return super.getType(stripQuery(uri));
     }
 
     @Override
-    public int delete(Uri uri, String selection, String[] selectionArgs) {
-        throw new UnsupportedOperationException();
+    public ParcelFileDescriptor openFile(@NonNull Uri uri, @NonNull String mode) throws FileNotFoundException {
+        return super.openFile(stripQuery(uri), mode);
+    }
+
+    @NonNull
+    @Override
+    public Cursor query(@NonNull Uri uri, @Nullable String[] projection, @Nullable String selection, @Nullable String[] selectionArgs, @Nullable String sortOrder) {
+        String[] columnNames = projection;
+        if (columnNames == null || columnNames.length == 0) {
+            columnNames = new String[] {
+                    OpenableColumns.DISPLAY_NAME, OpenableColumns.SIZE, "_data"
+            };
+        }
+
+        MatrixCursor newCursor = new MatrixCursor(columnNames);
+
+        File file = getFileForUri(uri);
+        if (file != null && file.exists()) {
+            String displayName = uri.getQueryParameter("display_name");
+            if (displayName == null) {
+                displayName = file.getName();
+            }
+
+            Object[] values = new Object[columnNames.length];
+            for (int i = 0; i < columnNames.length; i++) {
+                String column = columnNames[i];
+                if (OpenableColumns.DISPLAY_NAME.equals(column)) {
+                    values[i] = displayName;
+                } else if (OpenableColumns.SIZE.equals(column)) {
+                    values[i] = file.length();
+                } else if ("_data".equals(column)) {
+                    values[i] = file.getAbsolutePath();
+                } else {
+                    values[i] = null;
+                }
+            }
+            newCursor.addRow(values);
+        } else {
+            Timber.w("File not found for URI: %s", uri);
+            // If the file is not found, we still return the columns but no rows.
+            // Some apps might crash if we don't return the requested columns.
+        }
+
+        return newCursor;
+    }
+
+    private File getFileForUri(Uri uri) {
+        List<String> pathSegments = uri.getPathSegments();
+        if (pathSegments.size() < 2 || !"temp".equals(pathSegments.get(0))) {
+            return null;
+        }
+
+        String filename = pathSegments.get(1);
+        Context context = getContext();
+        if (context == null) return null;
+
+        return new File(getTempFileDirectory(context), filename);
+    }
+
+    @Override
+    public int delete(@NonNull Uri uri, @Nullable String selection, @Nullable String[] selectionArgs) {
+        return super.delete(stripQuery(uri), selection, selectionArgs);
+    }
+
+    private Uri stripQuery(Uri uri) {
+        if (uri.getQuery() == null) {
+            return uri;
+        }
+        return uri.buildUpon().query(null).build();
     }
 
     @Override
